% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGEDI4.R
\docType{package}
\name{GEDI4R}
\alias{GEDI4R}
\title{GEDI4R: A package for NASA's Global Ecosystem Dynamics Investigation (GEDI)
Level 4A Data Visualizing and Processing.}
\description{
The GEDI4R package provides functions for i) downloading, ii) reading, iii)
clipping, iv)visualizing and v) exporting GEDI data. The output of many
function is compatible with functions in the
\href{https://github.com/carlos-alberto-silva/rGEDI}{rGEDI} package. Use
\code{vignette("GEDI4R")} for an overview of the package functionalitis.
}
\note{
See the
\href{https://daac.ornl.gov/GEDI/guides/GEDI_L4A_AGB_Density.html}{user
guide} for more documentation on GEDI L4 data.
}
\author{
Elia Vangi
}
