% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedil4_sf.R
\docType{data}
\name{gedil4_sf}
\alias{gedil4_sf}
\title{Example of GEDI level 4A in sf format}
\format{
a \code{data.table} with 13153 observation and 13 variables.
\describe{see the Details section of \code{\link{l4_get}} for information on the
variables listed in this dataset}
}
\source{
The dataset was generated with functions provided in the
\code{GEDI4R} package and data downloaded from
\url{https://daac.ornl.gov/cgi-bin/dsviewer.pl?ds_id=1907}
}
\usage{
gedil4_sf
}
\description{
This dataset is generated by the function \code{l4_get} and clipped on the
Italian extent with \code{l4_clip(...,usebound=T)} using the extent of
\code{system.file("extdata","Italy.shp",package = "GEDI4R")} as bounding box.
}
\keyword{datasets}
