% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4_convert.R
\name{l4_convert}
\alias{l4_convert}
\title{Convert GEDI level 4A data to \code{sf} object and export it}
\usage{
l4_convert(gediL4, epsg = NULL, filename, return_path = F, ...)
}
\arguments{
\item{gediL4}{\code{data.table} obtained with the function
\code{l4_getmulti()} or \code{l4_clip}, or an \code{sf} object.}

\item{epsg}{Numeric: destination EPSG code. Default to NULL.}

\item{filename}{Character: path to data source name. Usually a file name with
extension. See \code{?st_write} for more details.}

\item{return_path}{Logical: if TRUE, the path of the saved file will be
returned, otherwise the function will return NULL. Default to FALSE.}

\item{...}{Others argument to pass to \code{st_write}, such as \code{driver}
or \code{append}}
}
\value{
The function is called for its side effects. It return \code{NULL}
unless \code{return_path=TRUE}.
}
\description{
Write GEDI level 4A data to disk as a vector file.
}
\details{
If a file extension will cause conflict with the driver, the
extension will be guessed from the file name. Note that in converting data
to ESRI Shapefile, columns names will be abbreviated with a warning.
}
\examples{

outdir = tempdir()
l4_zip <- system.file("extdata",
                     "GEDI04_A_2020036151358_O06515_02_T00198_02_002_01_V002.zip",
                     package="GEDI4R")
l4 <- unzip(l4_zip,exdir = outdir)
#get GEDI data
l4_data <- l4_getmulti(l4)
#convert to shp file
converted <- l4_convert(l4_data,epsg = 4326,filename=paste0(outdir,"/example.shp"),return_path = T,append=F)
example <- sf::read_sf(converted)
file.remove(list.files(outdir,pattern = "example",full.names = T))
}
\seealso{
\code{\link[sf]{st_write}}
}
