% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4_get.R
\name{l4_get}
\alias{l4_get}
\title{Get  AGBD data from a single path of GEDI Level 4 file}
\usage{
l4_get(gediL4_path, just_colnames = F, add_col = NULL, tct = NULL)
}
\arguments{
\item{gediL4_path}{Character path to GEDI level 4A h5 file.}

\item{just_colnames}{Logical: if TRUE, the function return the dataset
names in the h5 file. Useful to get names for the argument \code{add_col}.}

\item{add_col}{Character vector of extra datasets names to \code{cbind} to
the default output. Names already present will be dropped.}

\item{tct}{Numeric: tree cover threshold to filter GEDI data.}
}
\value{
an S4 object of class
\link[=data.table:data.table]{data.table}.
}
\description{
Get GEDI L4 data from h5 file format. See
\href{https://daac.ornl.gov/GEDI/guides/GEDI_L4A_AGB_Density.html}{here} for
more documentation on GEDI L4 data. This
function work as \code{l4_getmulti} but for a single file
}
\details{
Columns specification: Flag indicating likely invalid waveform
(1=valid,0=invalid).
\itemize{
\item \emph{date}: Date of acquisition in yyyy-mm-dd hh:mm:ss
\item \emph{beam}: Beam identifier
\item \emph{shot_number}: Shot number
\item \emph{degrade_flag }: Flag indicating
degraded state of pointing and/or positioning information
\item \emph{l4_quality_flag}: Flag identifying the most useful L4 data for
biomass predictions
\item \emph{l2_quality_flag}: Flag identifying the most
useful L2 data for biomass predictions
\item \emph{algorithm_run_flag}: The
L4A algorithm is run if this flag is set to 1. This flag selects data that
have sufficient waveform fidelity for AGBD estimation.
\item \emph{sensitivity}: Maxmimum canopy cover that can be penetrated
considering the SNR of the waveform
\item \emph{lat_lowestmode}: Latitude of center of lowest mode
\item \emph{lon_lowestmode}: Longitude of center of lowest mode
\item \emph{elev_lowestmode}: Elevation of center of lowest mode relative
to reference ellipsoid
\item \emph{tree_cover}: Tree cover in the year 2010, defined as canopy
closure for all vegetation taller than 5 m in height (Hansen et al., 2013)
and encoded as a percentage per output grid cell.
\item \emph{pft_class}: Plant Functional Type (PFT) derived from the MODIS
MCD12Q1 V006 product. Values follow the Land Cover Type 5 Classification
scheme.
\item \emph{agbd_se}: Aboveground biomass density (Mg/ha) prediction
standard error
\item \emph{agbd}: Aboveground biomass density (Mg/ha)
}
For downloading more detailed information on GEDI level 4A data (version 1)
see
\href{https://access.earthdata.nasa.gov/datasets/C2114031882-ORNL_CLOUD?action=index&controller=datasets&dataset_page_num=208&dataset_page_size=10&keyword=\%2A}{here}
Note that by default the function will dropp all footprint in which agbd<0.
These observations are considered sensor errors.
}
\examples{

#Specifying the path to GEDI level4A data (zip file)
outdir = tempdir()
l4_zip <- system.file("extdata",
                  "GEDI04_A_2020186052327_O08834_T03611_02_001_01.zip",
                  package="GEDI4R")

#Unzipping GEDI level 4A data
l4 <- unzip(l4_zip,exdir = outdir)
#dataset names
col_name <- l4_get(l4,just_colnames=T)
#get default data (see Details section)
l4_data <- l4_get(l4)
#filter footprint based on tree cover
l4_data_filter <- l4_get(l4,tct=10)
#add user-define dataset
l4_user_data <- l4_get(l4,add_col = col_name[2:3])
}
\seealso{
\code{\link[=l4_getmulti]{l4_getmulti}},
\code{\link[=l4_process]{l4_process}}
}
