% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4_getmulti.R
\name{l4_getmulti}
\alias{l4_getmulti}
\title{Get  AGBD data from a list of GEDI Level 4 file path}
\usage{
l4_getmulti(
  gediL4_path,
  just_colnames = F,
  add_col = NULL,
  tct = NULL,
  ncore = parallel::detectCores() - 1,
  merge = T,
  catch = T,
  source = F
)
}
\arguments{
\item{gediL4_path}{List or vector of character path to GEDI level4A h5 file.}

\item{just_colnames}{Logical: if TRUE, the function return the dataset
names in the h5 file. Useful to get names for the argument \code{add_col}.}

\item{add_col}{Character vector of extra datasets names to \code{cbind} to
the default output. Names already present will be dropped.}

\item{tct}{Numeric: tree cover threshold to filter GEDI data.}

\item{ncore}{Numeric: numbers of core to be used if the maximum core
available is less than the length of \code{gediL4_path} Default to the
number of cores available minus one.}

\item{merge}{Logical: if TRUE (default) the resulted list will be merged with
\code{rbind}. Ignored if \code{length(gediL4_path)==1}}

\item{catch}{Logical: if TRUE (default) files that would throw an error don't
interrupt the processing of subsequent files. For this file the function
will return NULL.Ignored if \code{length(gediL4_path)==1}}

\item{source}{Logical: if TRUE and \code{merge=TRUE} add a column with the
source path of each observation.Ignored if \code{length(gediL4_path)==1},
and if \code{merge=FALSE}}
}
\value{
A list of S4 object of class
\link[=data.table:data.table]{data.table} or a single object of
class
\link[=data.table:data.table]{data.table} if \code{merge=TRUE}.
}
\description{
Get GEDI L4 data from h5 file format. See
\href{https://daac.ornl.gov/GEDI/guides/GEDI_L4A_AGB_Density_V2.html}{here} for
more documentation on GEDI L4 data. The function use the
\link[snowfall:snowfall]{snowfall} package to get data in parallel. There
could be memory allocation problems for an high number of file path. This is
the core function of \code{\link[=l4_process]{l4_process}}, that manage the
whole workflow of reading, clipping and converting a list of h5 GEDI
file path.
}
\details{
Columns specification: Flag indicate likely invalid waveform
(1=valid,0=invalid).
\itemize{
\item \emph{date}: Date of acquisition in yyyy-mm-dd hh:mm:ss
\item \emph{beam}: Beam identifier
\item \emph{shot_number}: Shot number
\item \emph{degrade_flag }: Flag indicating
degraded state of pointing and/or positioning information
\item \emph{l4_quality_flag}: Flag identifying the most useful L4 data for
biomass predictions
\item \emph{l2_quality_flag}: Flag identifying the most
useful L2 data for biomass predictions
\item \emph{algorithm_run_flag}: The
L4A algorithm is run if this flag is set to 1. This flag selects data that
have sufficient waveform fidelity for AGBD estimation.
\item \emph{sensitivity}: Maxmimum canopy cover that can be penetrated
considering the SNR of the waveform
\item \emph{lat_lowestmode}: Latitude of center of lowest mode
\item \emph{lon_lowestmode}: Longitude of center of lowest mode
\item \emph{elev_lowestmode}: Elevation of center of lowest mode relative
to reference ellipsoid
\item \emph{tree_cover}: Tree cover in the year 2010, defined as canopy
closure for all vegetation taller than 5 m in height (Hansen et al., 2013)
and encoded as a percentage per output grid cell.
\item \emph{pft_class}: Plant Functional Type (PFT) derived from the MODIS
MCD12Q1 V006 product. Values follow the Land Cover Type 5 Classification
scheme.
\item \emph{agbd_se}: Aboveground biomass density (Mg/ha) prediction
standard error
\item \emph{agbd}: Aboveground biomass density (Mg/ha)
}
For DOWNLOADING more detailed information on GEDI Level 4A data (version 2)
see the
\href{https://daac.ornl.gov/daacdata/gedi/GEDI_L4A_AGB_Density_V2/comp/GEDI_L4A_V2_Product_Data_Dictionary.pdf}{product data dictionary}
Note that by default the function will dropp all footprint in which agbd<0.
These observations are considered sensor errors.
}
\examples{

Specifying the path to GEDI level4A data (zip file)
outdir = tempdir()
l4_zip <- system.file("extdata",
                      c("GEDI04_A_2020036151358_O06515_02_T00198_02_002_01_V002.zip",
                        "GEDI04_A_2021150031254_O13948_03_T06447_02_002_01_V002.zip"
                      ),
                      package="GEDI4R")
#Unzipping GEDI level4A data
l4 <- lapply(l4_zip,unzip,exdir = outdir)
#list all dataset in h5 file
l4_getmulti(l4[[1]],just_colnames=T)
#read and merge all file
l4_data <- l4_getmulti(l4,merge = T)
#return individual files as list
l4_data <- l4_getmulti(l4,merge = F)
#add user-defined dataset
col_name <- "delta_time"
l4_data <- l4_getmulti(l4,add_col = col_name,merge = T)
file.remove(list.files(outdir,full.names = T))
}
\seealso{
\code{\link[=l4_process]{l4_process}}
}
