% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4_plotprofile.R
\name{l4_plotprofile}
\alias{l4_plotprofile}
\title{Plot AGBD profile from GEDI level 4A}
\usage{
l4_plotprofile(gediL4, beam_id = "BEAM0000", tct = NULL)
}
\arguments{
\item{gediL4}{\code{data.table} or \code{sf_object}: obtained with the function
\code{l4_get}, \code{l4_getmulti(...,merge=TRUE)} or \code{l4_get}}

\item{beam_id}{Character: one of "BEAM0000" "BEAM0001" "BEAM0010" "BEAM0011"
"BEAM0101" "BEAM0110" "BEAM1000" "BEAM1011" or "all".}

\item{tct}{Numeric: tree cover threshold from 0 to 100 (in \%).}
}
\value{
Returns a \code{ggplot} object. See \link[ggplot2:ggplot]{ggplot2::ggplot} package.
}
\description{
The function plot the AGBD values against the along-track elevation profile.
The function is mainly used for plotting a single track instead of a full
area coverage of GEDI footprints.
}
\examples{
data("gedil4")
#plot first beam
l4_plotprofile(gedil4)
#plot all beams
l4_plotprofile(gedil4,beam_id="all")
}
\seealso{
\code{\link{l4_plotagb}}
}
