% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4_process.R
\name{l4_process}
\alias{l4_process}
\title{Process GEDI level 4 data.}
\usage{
l4_process(
  gediL4_path,
  nfile = NULL,
  clip = NULL,
  usegeometry = T,
  epsg = NULL,
  prefix = "",
  outdir = NULL,
  parallel = T,
  ncore = NULL,
  ...
)
}
\arguments{
\item{gediL4_path}{List or vector of character path to GEDI level4A h5 file.}

\item{nfile}{Numeric : number of file processed at time. It a affect the
numbers of total chunks in which files will be splitted. Default to NULL.}

\item{clip}{An object from which an extent can be obtained (xmin,ymin, xmax,
ymax). Currently could be a path to a shp or tif file, an object of class
\code{sf}, a Raster* object or a numeric vector of coordinates.}

\item{usegeometry}{Logical: should the points be clipped on the boundary of an
\code{sf_object} (or path from which an \code{sf_object} can be created). NULL and FALSE
will be interpreted as the same, in that case the extent of \code{clip}
will be used. Deafult to NULL.}

\item{epsg}{Numeric: destination EPSG code. Default to NULL.}

\item{prefix}{Character: prefix name for exported chunks. Default "".}

\item{outdir}{Character: directory in which to save ouput files. If don't
exist it will be created.}

\item{parallel}{Logical: if TRUE (default) the chain process are executetd in
parallel. In this case the function try to guess the best number of cores
to use based on the maximum available and the number of files to be
processed.}

\item{ncore}{Numeric: numbers of core to be used if the maximum core
available is less than the length of \code{gediL4_path} Default to the
number of cores available minus one.}

\item{...}{Other arguments to be passed to \code{l4_getmulti}.}
}
\value{
A vector of file paths inside \code{outdir}. Outputs will be written
as \code{outdir}/\code{prefix}_x_y.shp, where x and y are the first and last
file number in that chunk.
}
\description{
Perform the entire processing chain of extracting, clipping and exporting
GEDI data. See
\href{https://daac.ornl.gov/GEDI/guides/GEDI_L4A_AGB_Density.html}{here} for
more documentation on GEDI L4 data. This
is the main function of the package. It perform all operations in chunks
of files in parallel. It runs in order \code{l4_getmulti, l4_clip,
l4_convert} to each chunk.
}
\details{
The function use by default \code{usegeometry=TRUE} in \code{l4_clip}
and \code{append=FALSE} in \code{l4_convert}. Although they can be changed,
this is usually the desired behavior for the function. If
\code{parallel=TRUE}, the function will try to guess the best number of
cores to be used in \code{l4_getmulti} and in the chunks loop, based on the
maximum cores available and the length of \code{gediL4_path}. The user can
override the number of core used in \code{l4_getmulti} with \code{...}, by
specifying the argument \code{ncore}. This will also affect the number of
cores used to loop over chunks. Usually, the number of cores used by
default is the best option, modifying it can slow down the function.  See
\code{?l4_get} or \code{?l4_getmulti} for columns specification of the
output. For downloading more detailed information on GEDI Level 4A data
(version 1) see
\href{https://access.earthdata.nasa.gov/datasets/C2114031882-ORNL_CLOUD?action=index&controller=datasets&dataset_page_num=208&dataset_page_size=10&keyword=\%2A}{here}.
Note that by default, the function will drop all footprints in which agbd<0.
These observations are considered sensor errors.
}
\examples{

#Specifying the path to GEDI level4A data (zip file)
outdir  <-  tempdir()
l4_zip <- system.file("extdata",
                      c("GEDI04_A_2020186052327_O08834_T03611_02_001_01.zip",
                        "GEDI04_A_2020186065619_O08835_T00766_02_001_01.zip",
                        "GEDI04_A_2020187043633_O08849_T04437_02_001_01.zip",
                        "GEDI04_A_2020187060925_O08850_T01592_02_001_01.zip"
                      ),
                      package="GEDI4R")
#Unzipping GEDI level4A data
l4 <- lapply(l4_zip,unzip,exdir = outdir)
bound <- system.file("extdata","Italy.shp",package="GEDI4R")
#proces all files in chunk each of 2 files, in sequence
l4_data <- l4_process(l4,nfile=2,clip=bound,usegeometry=T,epsg=32632,prefix="block",outdir,parallel=F)
file.remove(l4_data)
#in parallel
l4_data <- l4_process(l4,nfile=2,clip=bound,usegeometry=T,epsg=32632,prefix="block",outdir,parallel=T)
file.remove(l4_data)
#override the default number of cores to be used
l4_data <- l4_process(l4,nfile=2,clip=bound,usegeometry=T,epsg=32632,prefix="block",outdir,parallel=T,ncore=2)
file.remove(l4_data)
}
\seealso{
\code{\link[=l4_getmulti]{l4_getmulti}},
\code{\link[=l4_clip]{l4_clip}},  \code{\link[=l4_convert]{l4_convert}}
}
